+++ pnet-data-api-java/src/main/java/pnet/data/api/person/PersonSettingsLinkDTO.java
@@ -1,0 +1,63 @@
+package pnet.data.api.person;
+
+import java.io.Serializable;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
+import pnet.data.api.util.WithTenant;
+
+/**
+ * Known settings of the user.
+ *
+ * @author HAM
+ */
+@ApiModel(description = "Holds the settings of the person per tenant.")
+public class PersonSettingsLinkDTO implements WithTenant, Serializable
+{
+    private static final long serialVersionUID = -8062595231537086982L;
+
+    @ApiModelProperty(notes = "The tenant these settings are defined for.")
+    private final String tenant;
+
+    @ApiModelProperty(notes = "The id of the preferred company.")
+    private final Integer preferredCompanyId;
+
+    @ApiModelProperty(notes = "The id of the preferred brand.")
+    private final String preferredBrandId;
+
+    public PersonSettingsLinkDTO(@JsonProperty("tenant") String tenant,
+        @JsonProperty("preferredCompanyId") Integer preferredCompanyId,
+        @JsonProperty("preferredBrandId") String preferredBrandId)
+    {
+        super();
+        this.tenant = tenant;
+        this.preferredCompanyId = preferredCompanyId;
+        this.preferredBrandId = preferredBrandId;
+    }
+
+    @Override
+    public String getTenant()
+    {
+        return tenant;
+    }
+
+    public Integer getPreferredCompanyId()
+    {
+        return preferredCompanyId;
+    }
+
+    public String getPreferredBrandId()
+    {
+        return preferredBrandId;
+    }
+
+    @Override
+    public String toString()
+    {
+        return String
+            .format("PersonSettingsLinkDTO [tenant=%s, preferredCompanyId=%s, preferredBrandId=%s]", tenant,
+                preferredCompanyId, preferredBrandId);
+    }
+}